pyproject.toml
[tool.poetry]
name = "s3-bench"
version = "0.1.0"
description = ""
authors = ["Alexander Dich Jensen <dichjensen@gmail.com>"]
readme = "README.md"
packages = [{include = "s3_bench"}]

[tool.poetry.dependencies]
python = "~3.11"
boto3 = "^1.28.62"
pandas = "^2.1.1"
tqdm = "^4.66.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
flake8 = "^6.1.0"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120